[inductor] Move debugging code to torchinductor.debug (#996)
